src/requirements.txt
functions-framework
google-auth
google-cloud-aiplatform==1.25.0
google-cloud-bigquery
google-cloud-logging
google-cloud-storage
google-cloud-vision